requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
altgraph==0.17.4
    # via modulegraph
kybra==0.5.2
    # via -r requirements.in
modulegraph==0.19.3
    # via kybra

# The following packages are considered to be unsafe in a requirements file:
# setuptools

# AI
openai-whisper
flask
ffmpeg-python
